+++ src/rules/property-whitelist/__tests__/index.js
@@ -64,6 +64,12 @@ testRule(rule, {
     code: "a { background-color: pink; }",
   }, {
     code: "a { background-image: none; }",
+  }, {
+    code: "a { $scss: 0; }",
+  }, {
+    code: "a { @less: 0; }",
+  }, {
+    code: "a { --custom-property: 0; }",
   } ],
 
   reject: [{
@@ -72,4 +78,20 @@ testRule(rule, {
     line: 1,
     column: 5,
   }],
+})
+
+testRule(rule, {
+  ruleName,
+
+  config: [[
+    "/margin/",
+  ]],
+
+  accept: [ {
+    code: "a { $padding: 0; }",
+  }, {
+    code: "a { @padding: 0; }",
+  }, {
+    code: "a { --padding: 0; }",
+  } ],
 })